Generators for material blocks
